// services/document-evidence/Generated/frontend-app/src/generated-ts/documents/types/UpdateDocumentRequest.generated.ts
// ------------------------------------------------------------------------------
// GENERATED FILE - source: documents/types.json
// DO NOT EDIT MANUALLY.
// ------------------------------------------------------------------------------

export interface UpdateDocumentRequest {
  "currentRevisionNumber": number;
  "displayName": string;
  "isRevoked": boolean;
  "mediaType": string;
  "sourceKind": string;
  "state": string;
  "workspaceId": string;
}
export const UpdateDocumentRequestTypeKey = "LogicLens.DocumentEvidence.Generated.Api.Contracts.UpdateDocumentRequest" as const;